added EMS

+++ DotNetCore/DotNet Assignment/Day2/Services/ProductService.cs
@@ -4,6 +4,8 @@ using Day2.DTO;
 using Day2.Interfaces;
 using Day2.Models;
 using Microsoft.EntityFrameworkCore;
+using System.Collections;
+using System.Data;
 
 namespace Day2.Services
 {
@@ -33,28 +35,43 @@ namespace Day2.Services
                                    .AsNoTracking().ToList();
             return _mapper.Map<IEnumerable<ProductReadDto>>(products);
         }
-        public ProductReadDto Add(ProductCreateDto dto)
+        public ProductReadDto Add(ProductCreateDto dto,Guid vendorId)
         {
             var product = _mapper.Map<Product>(dto);
+            product.VendorId = vendorId;
             _context.Products.Add(product);
             _context.SaveChanges();
             return _mapper.Map<ProductReadDto>(product);
         }
-        public bool Update(int id, ProductUpdateDto dto)
+        public bool Update(int id, ProductUpdateDto dto,Guid userid, string role)
         {
             var exist = _context.Products.Find(id);
             if (exist == null) return false;
+            if(role=="Admin")
+            {
+                _mapper.Map(dto, exist);
+                _context.SaveChanges();
+                return true;
+            }
+            if(role=="Vendor" && exist.VendorId != userid)
+            {
+                return false;
+            }
             _mapper.Map(dto, exist);
             _context.SaveChanges();
             return true;
         }
-        public bool Delete(int id)
+        public bool Delete(int id,Guid userid, string role)
         {
             var product = _context.Products.Find(id);
             if (product == null) return false;
-            _context.Products.Remove(product);
-            _context.SaveChanges();
-            return true;
+            if (role == "Admin")
+            {
+                _context.Products.Remove(product);
+                _context.SaveChanges();
+                return true;
+            }
+            return false;
         }
     }
 }
